backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
openai>=1.30.0
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
httplib2>=0.20.0
